komentarai

+++ app/Http/Controllers/SkelbimaiController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Ad;
 use App\Category;
 use Illuminate\Http\Request;
+use Gate;
+use Illuminate\Support\Facades\Auth;
 use File;
 
 class SkelbimaiController extends Controller
@@ -32,7 +34,6 @@ class SkelbimaiController extends Controller
 
         $path=$request->file('nuotrauka')->store('public/images');
         $filename=str_replace('public/',"", $path);
-
         $ad = Ad::create([
             'catid' => request('kategorija'),
             'pavadinimas' => request('pavadinimas'), // pavadinimas db laukelis, o key input name
@@ -41,9 +42,13 @@ class SkelbimaiController extends Controller
             'email' => request('email'),
             'phone' => request('phone'),
             'location' => request('vieta'),
-            'nuotrauka' => $filename
+            'nuotrauka' => $filename,
+            'user' => auth()->user()->id
+//            'user' => Auth::id()
+
 
         ]);
+
            return redirect('/ads');
 //        $cat = request(['kategorija','pavadinimas','aprasymas','kaina','email','vieta','phone']);
 //        dd($cat);
@@ -54,6 +59,7 @@ class SkelbimaiController extends Controller
         $category = Category::all();
         return view('skelbimai.pages.ads', compact('ads'),compact('category'));
 
+
     }
     public function adDelete(Ad $ad)
     {
@@ -62,41 +68,43 @@ class SkelbimaiController extends Controller
     }
     public function adData(Ad $ad)   {
 
-        return view('skelbimai.pages.addata', compact('ad'));
+            return view('skelbimai.pages.addata', compact('ad'));
     }
     public function adUpdate(Ad $ad,Request $request){
-        $validatedData = $request->validate([
-            'pavadinimas' => 'required',
-            'aprasymas' => 'required',
-            'kaina' => 'required',
-            'email' => 'required',
-            'vieta' => 'required',
-            'nuotrauka' => 'mimes:jpeg,jpg,png,gif|required|max:10000'
-        ]);
-        $path=$request->file('nuotrauka')->store('public/images');
-        $filename=str_replace('public/',"", $path);
+        if (Gate::allows('adUpdate',$ad)) {
+            $validatedData = $request->validate([
+                'pavadinimas' => 'required',
+                'aprasymas' => 'required',
+                'kaina' => 'required',
+                'email' => 'required',
+                'vieta' => 'required',
+                'nuotrauka' => 'mimes:jpeg,jpg,png,gif|required|max:10000'
+            ]);
+            $path = $request->file('nuotrauka')->store('public/images');
+            $filename = str_replace('public/', "", $path);
 
-        Ad::where ('id', request('id'))->
-        update(['pavadinimas' => request('pavadinimas'),
+            Ad::where('id', request('id'))->
+            update(['pavadinimas' => request('pavadinimas'),
                 'aprasymas' => request('aprasymas'),
                 'price' => request('kaina'),
                 'email' => request('email'),
                 'phone' => request('phone'),
                 'location' => request('vieta'),
                 'nuotrauka' => $filename
-        ]);
-        //patikrina ar postas turi nuotrauka ir istrina sena.
-        if($request->hasFile('nuotrauka'))
-        {
-            File::delete('../storage/app/public/'.$ad->nuotrauka);
-            $path=$request->file('nuotrauka')->store('public/images');
-            $filename=str_replace('public/',"", $path);
-                Ad::where('id',$ad->id)->update([
-                   'nuotrauka' => $filename
+            ]);
+            //patikrina ar postas turi nuotrauka ir istrina sena.
+            if ($request->hasFile('nuotrauka')) {
+                File::delete('../storage/app/public/' . $ad->nuotrauka);
+                $path = $request->file('nuotrauka')->store('public/images');
+                $filename = str_replace('public/', "", $path);
+                Ad::where('id', $ad->id)->update([
+                    'nuotrauka' => $filename
                 ]);
+            }
+
         }
 
-        return redirect('/ads');
+        return redirect('/error');
     }
 }
